# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Acompanhamento Processual",
    "description": "API para gestão e acompanhamento de processos jurídicos com sistema de alertas automáticos",
    "version": "1.0.0",
    "contact": {
      "name": "Sistema Processual",
      "email": "contato@sistemaprocessual.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Servidor de Desenvolvimento"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "role": { "type": "string", "enum": ["admin", "user"] },
          "ativo": { "type": "boolean" }
        }
      },
      "Processo": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "numero": { "type": "string" },
          "classe": { "type": "string" },
          "assunto": { "type": "string" },
          "tribunal": { "type": "string" },
          "comarca": { "type": "string" },
          "status": { "type": "string", "enum": ["ativo", "arquivado", "suspenso"] },
          "dataDistribuicao": { "type": "string", "format": "date" },
          "dataSentenca": { "type": "string", "format": "date" },
          "prazoRecurso": { "type": "string", "format": "date" },
          "prazoEmbargos": { "type": "string", "format": "date" },
          "proximaAudiencia": { "type": "string", "format": "date" },
          "observacoes": { "type": "string" }
        }
      },
      "Alert": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "tipo": { "type": "string", "enum": ["audiencia", "prazo_recurso", "prazo_embargos", "despacho"] },
          "titulo": { "type": "string" },
          "mensagem": { "type": "string" },
          "dataVencimento": { "type": "string", "format": "date" },
          "prioridade": { "type": "string", "enum": ["baixa", "media", "alta", "urgente"] },
          "lido": { "type": "boolean" }
        }
      }
    }
  },
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Registra um novo usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nome": { "type": "string", "minLength": 2 },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 6 },
                  "role": { "type": "string", "enum": ["admin", "user"] }
                },
                "required": ["nome", "email", "password"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Autentica um usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login realizado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "token": { "type": "string" },
                    "user": { "$ref": "#/components/schemas/User" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/processos": {
      "get": {
        "tags": ["Processos"],
        "summary": "Lista processos do usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de processos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "processos": { "type": "array", "items": { "$ref": "#/components/schemas/Processo" } },
                    "pagination": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Processos"],
        "summary": "Cria um novo processo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "numero": { "type": "string", "minLength": 10 },
                  "classe": { "type": "string", "minLength": 2 },
                  "assunto": { "type": "string" },
                  "tribunal": { "type": "string" },
                  "comarca": { "type": "string" },
                  "dataDistribuicao": { "type": "string", "format": "date" },
                  "dataSentenca": { "type": "string", "format": "date" },
                  "proximaAudiencia": { "type": "string", "format": "date" },
                  "observacoes": { "type": "string" }
                },
                "required": ["numero", "classe"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Processo criado com sucesso"
          }
        }
      }
    },
    "/alerts": {
      "get": {
        "tags": ["Alertas"],
        "summary": "Lista alertas do usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer" } },
          { "name": "tipo", "in": "query", "schema": { "type": "string" } },
          { "name": "lido", "in": "query", "schema": { "type": "boolean" } },
          { "name": "prioridade", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Lista de alertas",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "alertas": { "type": "array", "items": { "$ref": "#/components/schemas/Alert" } },
                    "pagination": { "type": "object" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/external/processos/{numero}": {
      "get": {
        "tags": ["API Externa"],
        "summary": "Consulta processo na API externa (DataJud)",
        "parameters": [
          {
            "name": "numero",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Número do processo"
          }
        ],
        "responses": {
          "200": {
            "description": "Processo encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "processo": { "type": "object" },
                    "source": { "type": "string" }
                  }
                }
              }
            }
          },
          "503": {
            "description": "Serviço externo não configurado"
          }
        }
      }
    }
  }
}